+++ R/plot.estimate_density.R
@@ -1,6 +1,6 @@
 #' @importFrom insight clean_parameters
 #' @export
-data_plot.estimate_density <- function(x, data = NULL, ...) {
+data_plot.estimate_density <- function(x, data = NULL, centrality = "median", ci = 0.95, ...) {
   dataplot <- x
 
   if (!"Parameter" %in% names(dataplot)) {
@@ -17,7 +17,36 @@ data_plot.estimate_density <- function(x, data = NULL, ...) {
   dataplot$Parameter <- factor(dataplot$Parameter)
   dataplot$Parameter <- factor(dataplot$Parameter, levels = rev(levels(dataplot$Parameter)))
 
+  # summary
+  split_columns <- intersect(c("Parameter", "Effects", "Component"), colnames(dataplot))
+  datasplit <- split(dataplot, dataplot[split_columns])
+  summary <- do.call(rbind, .compact_list(lapply(datasplit, function(i) {
+    if (length(i$x) > 0) {
+      Estimate <- as.numeric(bayestestR::point_estimate(i$x, centrality = centrality))
+      CI <- as.numeric(bayestestR::ci(i$x, ci = ci))
+      out <- data.frame(
+        Parameter = unique(i$Parameter),
+        x = Estimate,
+        CI_low = CI[2],
+        CI_high = CI[3],
+        stringsAsFactors = FALSE
+      )
+      if ("Effects" %in% colnames(i)) {
+        out$Effects <- unique(i$Effects)
+      }
+      if ("Component" %in% colnames(i)) {
+        out$Component <- unique(i$Component)
+      }
+    } else {
+      out <- NULL
+    }
+    out
+  })))
 
+  summary$Parameter <- factor(summary$Parameter)
+  summary$Parameter <- factor(summary$Parameter, levels = levels(dataplot$Parameter))
+
+  attr(dataplot, "summary") <- summary
   attr(dataplot, "info") <- list("xlab" = "Values",
                                  "ylab" = "Density",
                                  "legend_fill" = "Parameter",
@@ -37,14 +66,16 @@ data_plot.estimate_density <- function(x, data = NULL, ...) {
 #'
 #' The \code{plot()} method for the \code{bayestestR::estimate_density()} function.
 #'
-#' @param stack Logical, if \code{TRUE}, densities are plotted as stacked lines.
-#'   Else, densities are plotted for each parameter among each other.
-#' @param priors Logical, if \code{TRUE}, prior distributions are simulated
-#'   (using \code{\link[bayestestR]{simulate_prior}}) and added to the plot.
+#' @param stack Logical, if \code{TRUE}, densities are plotted as stacked lines. Else, densities are plotted for each parameter among each other.
+#' @param priors Logical, if \code{TRUE}, prior distributions are simulated (using \code{\link[bayestestR:simulate_prior]{simulate_prior()}}) and added to the plot.
 #' @param priors_alpha Alpha value of the prior distributions.
+#' @param posteriors_alpha Alpha value of the posterior distributions.
+#' @param centrality The point-estimate (centrality index) to compute. May be \code{"median"}, \code{"mean"} or \code{"MAP"}.
+#' @param ci Value of probability of the CI (between 0 and 1) to be estimated. Default to .95.
 #' @inheritParams data_plot
 #' @inheritParams plot.see_bayesfactor_parameters
 #' @inheritParams plot.see_cluster_analysis
+#' @inheritParams plot.see_check_normality
 #'
 #' @return A ggplot2-object.
 #'
@@ -60,7 +91,7 @@ data_plot.estimate_density <- function(x, data = NULL, ...) {
 #' @importFrom rlang .data
 #' @importFrom ggridges geom_ridgeline
 #' @export
-plot.see_estimate_density <- function(x, stack = TRUE, show_intercept = FALSE, n_columns = 1, priors = FALSE, priors_alpha = .4, size = .9, ...) {
+plot.see_estimate_density <- function(x, stack = TRUE, show_intercept = FALSE, n_columns = 1, priors = FALSE, priors_alpha = .4, posteriors_alpha = 0.7, size_line = .9, size_point = 2, centrality = "median", ci = 0.95, ...) {
   # save model for later use
   model <- tryCatch(
     {
@@ -74,7 +105,7 @@ plot.see_estimate_density <- function(x, stack = TRUE, show_intercept = FALSE, n
 
 
   if (!"data_plot" %in% class(x)) {
-    x <- data_plot(x, data = model, ...)
+    x <- data_plot(x, data = model, centrality = centrality, ci = ci, ...)
   }
 
   if ((!"Effects" %in% names(x) || length(unique(x$Effects)) <= 1) &&
@@ -87,24 +118,12 @@ plot.see_estimate_density <- function(x, stack = TRUE, show_intercept = FALSE, n
   x <- .remove_intercept(x, show_intercept = show_intercept)
 
   if (stack == TRUE) {
-    p <- ggplot(
-      x,
-      aes(
-        x = .data$x,
-        y = .data$y,
-        color = .data$Parameter
-      )) +
-      geom_line(size = size) +
+    p <- ggplot(x, aes(x = .data$x, y = .data$y, color = .data$Parameter)) +
+      geom_line(size = size_line) +
       add_plot_attributes(x) +
       scale_color_flat(labels = labels)
   } else {
-    p <- ggplot(
-      x,
-      aes(
-        x = .data$x,
-        y = .data$Parameter,
-        height = .data$y
-      ))
+    p <- ggplot(x, aes(x = .data$x, y = .data$Parameter, height = .data$y))
 
     # add prior layer
     if (priors) {
@@ -112,13 +131,28 @@ plot.see_estimate_density <- function(x, stack = TRUE, show_intercept = FALSE, n
         .add_prior_layer_ridgeline(
           model,
           show_intercept = show_intercept,
-          priors_alpha = priors_alpha
+          priors_alpha = priors_alpha,
+          show_ridge_line = FALSE
         ) +
-        ggridges::geom_ridgeline(aes(fill = "Posterior"), alpha = .7) +
-        scale_fill_flat(reverse = TRUE)
+        ggridges::geom_ridgeline(aes(fill = "Posterior"), alpha = posteriors_alpha, color = NA) +
+        guides(color = "none") +
+        scale_fill_flat(reverse = TRUE) +
+        scale_colour_flat(reverse = TRUE)
     } else {
-      p <- p + ggridges::geom_ridgeline()
+      p <- p +
+        ggridges::geom_ridgeline(aes(fill = "Posterior"), alpha = posteriors_alpha, color = NA) +
+        guides(fill = "none", color = "none") +
+        scale_fill_manual(values = unname(social_colors("grey"))) +
+        scale_color_manual(values = unname(social_colors("grey")))
     }
+
+    summary <- attributes(x)$summary
+    summary <- .remove_intercept(summary, show_intercept = show_intercept)
+    summary$y <- NA
+
+    p <- p +
+      geom_errorbarh(data = summary, mapping = aes(xmin = .data$CI_low, xmax = .data$CI_high, color = "Posterior"), size = size_line) +
+      geom_point(data = summary, mapping = aes(x = .data$x, color = "Posterior"), size = size_point, fill = "white", shape = 21)
 
     p <- p + add_plot_attributes(x)
   }
@@ -162,27 +196,15 @@ data_plot.estimate_density_df <- data_plot.estimate_density
 #' @importFrom ggridges geom_ridgeline
 #' @importFrom stats setNames
 #' @export
-plot.see_estimate_density_df <- function(x, stack = TRUE, n_columns = 1, size = .9, ...) {
+plot.see_estimate_density_df <- function(x, stack = TRUE, n_columns = 1, size_line = .9, ...) {
   x$Parameter <- factor(x$Parameter, levels = rev(unique(x$Parameter)))
   labels <- stats::setNames(levels(x$Parameter), levels(x$Parameter))
 
   if (stack == TRUE) {
-    p <- ggplot(
-      x,
-      aes(
-        x = .data$x,
-        y = .data$y,
-        color = .data$Parameter
-      )) +
-      geom_line(size = size)
+    p <- ggplot(x, aes(x = .data$x, y = .data$y, color = .data$Parameter)) +
+      geom_line(size = size_line)
   } else {
-    p <- ggplot(
-      x,
-      aes(
-        x = .data$x,
-        y = .data$Parameter,
-        height = .data$y
-      )) +
+    p <- ggplot(x, aes(x = .data$x, y = .data$Parameter, height = .data$y)) +
       ggridges::geom_ridgeline()
   }